use millisecond precision for time properties

+++ lib/mixpanel-ruby/events.rb
@@ -55,7 +55,7 @@ module Mixpanel
       properties = {
         'distinct_id' => distinct_id,
         'token' => @token,
-        'time' => Time.now.to_i,
+        'time' => Time.now.to_f,
         'mp_lib' => 'ruby',
         '$lib_version' => Mixpanel::VERSION,
       }.merge(properties)
@@ -103,7 +103,7 @@ module Mixpanel
       properties = {
         'distinct_id' => distinct_id,
         'token' => @token,
-        'time' => Time.now.to_i,
+        'time' => Time.now.to_f,
         'mp_lib' => 'ruby',
         '$lib_version' => Mixpanel::VERSION,
       }.merge(properties)
